DB에 유저 정보 저장 기능 구현
DB에 유저 정보를 저장하고 리스트로 목록을 불러오는 기능을 구현 하였습니다.
- User 클래스 Entity 선언
- UserRepository 추가, save, findAll 메서드로 저장 및 출력
