+++ src/main/java/com/codesquad/qna/domain/UserRepository.java
@@ -1,0 +1,6 @@
+package com.codesquad.qna.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
